+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- add `@className` decorator
+- fix build_runner error
+
 ## 0.1.0
 
 - refactor to use factory pattern